Moved package mapper from service to application and fixed some grammatical mistakes.
