+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/apiservice/UserApiServiceImpl.java
@@ -2,9 +2,9 @@ package com.softserveinc.ita.homeproject.application.apiservice;
 
 import com.softserveinc.ita.homeproject.api.UsersApiService;
 import com.softserveinc.ita.homeproject.homeservice.dto.UserDto;
-import com.softserveinc.ita.homeproject.homeservice.mapperViewToDto.CreateUserDtoMapper;
-import com.softserveinc.ita.homeproject.homeservice.mapperViewToDto.ReadUserDtoMapper;
-import com.softserveinc.ita.homeproject.homeservice.mapperViewToDto.UpdateUserDtoMapper;
+import com.softserveinc.ita.homeproject.application.mapper.CreateUserDtoMapper;
+import com.softserveinc.ita.homeproject.application.mapper.ReadUserDtoMapper;
+import com.softserveinc.ita.homeproject.application.mapper.UpdateUserDtoMapper;
 import com.softserveinc.ita.homeproject.homeservice.service.UserService;
 import com.softserveinc.ita.homeproject.model.CreateUser;
 import com.softserveinc.ita.homeproject.model.ReadUser;
