+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/mapper/DtoToViewMapper.java
@@ -1,4 +1,4 @@
-package com.softserveinc.ita.homeproject.homeservice.mapperViewToDto;
+package com.softserveinc.ita.homeproject.application.mapper;
 
 import com.softserveinc.ita.homeproject.homeservice.dto.BaseDto;
