+++ home-data/src/main/java/com/softserveinc/ita/homeproject/homedata/entity/Permission.java
@@ -1,18 +1,13 @@
 package com.softserveinc.ita.homeproject.homedata.entity;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 import javax.persistence.Entity;
 import javax.persistence.Column;
 import javax.persistence.ManyToMany;
 import java.util.Set;
 
-@Getter
-@Setter
+@Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
