+++ home-data/src/main/java/com/softserveinc/ita/homeproject/homedata/entity/Role.java
@@ -1,10 +1,6 @@
 package com.softserveinc.ita.homeproject.homedata.entity;
 
-import lombok.Getter;
-import lombok.Setter;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
+import lombok.*;
 
 import javax.persistence.Entity;
 import javax.persistence.Column;
@@ -14,8 +10,7 @@ import javax.persistence.JoinColumn;
 import java.util.Set;
 
 @Entity
-@Getter
-@Setter
+@Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
