+++ home-data/src/main/java/com/softserveinc/ita/homeproject/homedata/entity/User.java
@@ -1,10 +1,6 @@
 package com.softserveinc.ita.homeproject.homedata.entity;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 import javax.persistence.Entity;
 import javax.persistence.Column;
@@ -15,8 +11,7 @@ import javax.persistence.JoinColumn;
 import java.time.LocalDateTime;
 import java.util.Set;
 
-@Getter
-@Setter
+@Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
