+++ pom.xml
@@ -29,7 +29,7 @@
         <spring.boot.version>2.3.4.RELEASE</spring.boot.version>
         <maven.compiler.plugin.version>3.8.1</maven.compiler.plugin.version>
         <lombok.version>1.18.16</lombok.version>
-        <mupstruct.version>1.4.1.Final</mupstruct.version>
+        <mapstruct.version>1.4.1.Final</mapstruct.version>
     </properties>
 
     <dependencyManagement>
@@ -58,12 +58,12 @@
         <dependency>
             <groupId>org.mapstruct</groupId>
             <artifactId>mapstruct</artifactId>
-            <version>${mupstruct.version}</version>
+            <version>${mapstruct.version}</version>
         </dependency>
         <dependency>
             <groupId>org.mapstruct</groupId>
             <artifactId>mapstruct-processor</artifactId>
-            <version>${mupstruct.version}</version>
+            <version>${mapstruct.version}</version>
         </dependency>
     </dependencies>
